Time: 4 ms (100.00%), Space: 61.2 MB (77.55%) - LeetHub

+++ 2962-count-subarrays-where-max-element-appears-at-least-k-times/2962-count-subarrays-where-max-element-appears-at-least-k-times.java
@@ -1,6 +1,6 @@
 class Solution {
     public long countSubarrays(int[] nums, int k) {
-        long max=0;
+        int max=0;
         int i=0;
         int j=0;
         int n=nums.length;
